# docs/swagger.yaml
basePath: /
definitions:
...
  authorization.MemberRoleView:
    properties:
      code:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  authorization.PermissionView:
    properties:
      action:
        type: string
      id:
        type: string
      name:
        type: string
      permission_key:
        type: string
      resource:
        type: string
    type: object
  authorization.TenantRole:
    properties:
      code:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      version:
        type: integer
    type: object
  authorization.TenantRolePage:
    properties:
      items:
        items:
          $ref: '#/definitions/authorization.TenantRole'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  httptransport.DeleteTenantRoleRequest:
    properties:
      id:
        type: string
      version:
        type: integer
    required:
    - id
    - version
    type: object
...
  httptransport.Response:
    properties:
      body: {}
      code:
        type: integer
      message:
        type: string
      request_id:
        type: string
    type: object
...
  httptransport.TenantRoleIDRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  httptransport.TenantRolePageRequest:
    properties:
      created_at_from:
        type: string
      created_at_to:
        type: string
      ids:
        items:
          type: string
        type: array
      keyword:
        type: string
      page:
        type: integer
      page_size:
        type: integer
      statuses:
        items:
          type: string
        type: array
    type: object
...
  httptransport.UpdateTenantRoleRequest:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: string
      version:
        type: integer
    required:
    - id
    - name
    - status
    - version
    type: object
...
  httptransport.createTenantRoleRequest:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      permission_ids:
        items:
          type: string
        type: array
    required:
    - code
    - name
    type: object
  httptransport.effectivePermissionsRequest:
    properties:
      membership_id:
        type: string
    type: object
  httptransport.memberRolesRequest:
    properties:
      membership_id:
        type: string
    required:
    - membership_id
    type: object
...
  httptransport.setAdministratorRequest:
    properties:
      enabled:
        type: boolean
      membership_id:
        type: string
      tenant_id:
        type: string
    required:
    - membership_id
    - tenant_id
    type: object
  httptransport.setMemberRolesRequest:
    properties:
      membership_id:
        type: string
      role_ids:
        items:
          type: string
        type: array
    required:
    - membership_id
    type: object
  httptransport.setRolePermissionsRequest:
    properties:
      permission_ids:
        items:
          type: string
        type: array
      role_id:
        type: string
      version:
        type: integer
    required:
    - role_id
    - version
    type: object
  httptransport.setTenantPermissionsRequest:
    properties:
      permission_ids:
        items:
          type: string
        type: array
      tenant_id:
        type: string
    required:
    - tenant_id
    type: object
...
info:
  contact: {}
  description: 'Production-oriented Go Web API scaffold. Application error codes:
    0 success; 10000-19999 common/input; 20000-29999 authentication/authorization;
    30000-39999 business; 50000-59999 infrastructure.'
  title: Go API Template
  version: "1.0"
paths:
...
  /api/v1/tenant-authorization/administrators/set:
    post:
      consumes:
      - application/json
      parameters:
      - description: Administrator
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.setAdministratorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Add or remove a tenant administrator
      tags:
      - tenant-authorization
  /api/v1/tenant-authorization/effective-permissions:
    post:
      consumes:
      - application/json
      parameters:
      - description: Member
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.effectivePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  items:
                    type: string
                  type: array
              type: object
      security:
      - Bearer: []
      summary: List effective permission IDs for a tenant member
      tags:
      - tenant-authorization
  /api/v1/tenant-authorization/permissions/set:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant permissions
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.setTenantPermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Replace a tenant's permission ceiling
      tags:
      - tenant-authorization
...
  /api/v1/tenant-members/roles/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: Member
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.memberRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/authorization.MemberRoleView'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: List displayable roles assigned to a tenant member
      tags:
      - tenant-authorization
  /api/v1/tenant-members/roles/set:
    post:
      consumes:
      - application/json
      parameters:
      - description: Member roles
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.setMemberRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Replace a tenant member's roles
      tags:
      - tenant-authorization
...
  /api/v1/tenant-roles/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.createTenantRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/authorization.TenantRole'
              type: object
      security:
      - Bearer: []
      summary: Create a tenant role from the caller's effective permissions
      tags:
      - tenant-authorization
  /api/v1/tenant-roles/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.DeleteTenantRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Delete a tenant role and its assignments
      tags:
      - tenant-authorization
  /api/v1/tenant-roles/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: Role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.TenantRoleIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/authorization.TenantRole'
              type: object
      security:
      - Bearer: []
      summary: Get a tenant role
      tags:
      - tenant-authorization
  /api/v1/tenant-roles/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: Filters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.TenantRolePageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/authorization.TenantRolePage'
              type: object
      security:
      - Bearer: []
      summary: Page tenant roles
      tags:
      - tenant-authorization
  /api/v1/tenant-roles/permissions/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: Role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.TenantRoleIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/authorization.PermissionView'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: List displayable permissions assigned to a tenant role
      tags:
      - tenant-authorization
  /api/v1/tenant-roles/permissions/set:
    post:
      consumes:
      - application/json
      parameters:
      - description: Role permissions
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.setRolePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Replace role permissions with optimistic locking
      tags:
      - tenant-authorization
  /api/v1/tenant-roles/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: Role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.UpdateTenantRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/authorization.TenantRole'
              type: object
      security:
      - Bearer: []
      summary: Update a tenant role
      tags:
      - tenant-authorization
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Enter "Bearer {JWT}".
    in: header
    name: Authorization
    type: apiKey
  PSK:
    description: Enter "PSK {shared-key}" for routes configured with PSK authentication.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"